test: keep one Swift E2E suite per file
Separate alias and deprecated command specs into their own files so the
source layout matches the suite identity. This keeps the reference docs
and recording metadata easy to reason about as the alias coverage grows.
